+++ library/Morloc/Pretty.hs
@@ -19,6 +19,16 @@ import Morloc.Namespace
 import qualified Data.Map as Map
 import qualified Data.Set as Set
 
+instance Pretty Symbol where
+  pretty (TypeSymbol x) = viaShow x
+  pretty (TermSymbol x) = viaShow x
+
+instance Pretty AliasedSymbol where
+  pretty (AliasedType x alias) = pretty (AliasedTerm x alias)
+  pretty (AliasedTerm x alias)
+    | x == alias = pretty x
+    | otherwise = pretty x <+> "as" <+> pretty alias
+
 instance Pretty MVar where
   pretty = pretty . unMVar
 
@@ -47,15 +57,6 @@ instance Pretty Source where
     <> maybe "" (\ path -> " from" <+> dquotes (pretty path)) (srcPath s)
     <+> dquotes (pretty (srcName s))
     <+> "as" <+> pretty (srcAlias s) <> maybe "" (\t -> ":" <> pretty t) (srcLabel s)
-
-forallVars :: TypeU -> [Doc ann]
-forallVars (ForallU (TV _ v) t) = pretty v : forallVars t
-forallVars _ = []
-
-forallBlock :: TypeU -> Doc ann
-forallBlock (ForallU _ t) = forallBlock t
-forallBlock t = pretty t
-
 
 instance Pretty Type where
   pretty (UnkT (TV _ v)) = "*" <> pretty v
@@ -96,20 +97,26 @@ instance Pretty Constraint where
   pretty (Con x) = pretty x
 
 instance Pretty TypeU where
-  pretty (ExistU v ts ds)
-    = angles $ pretty v
-    <> list (map pretty ts)
-    <> list (map pretty ds)
-  pretty t@(ForallU _ _) =
-    parens $ "forall" <+> hsep (forallVars t) <+> "." <+> forallBlock t
-  pretty (VarU (TV _ "Unit")) = "()"
-  pretty (VarU v) = pretty v
-  pretty (FunU [] t) = parens $ "<MISSING> -> " <> pretty t
-  pretty (FunU ts t) = encloseSep "(" ")" " -> " (map pretty (ts <> [t]))
-  pretty (AppU t ts) = parens . hsep $ map pretty (t:ts)
-  pretty (NamU o n ps rs)
-      = block 4 (viaShow o <+> pretty n <> encloseSep "<" ">" "," (map pretty ps))
-                (vsep [pretty k <+> "::" <+> pretty x | (k, x) <- rs])
+  pretty (FunU [] t) = "<MISSING> -> " <> prettyTypeU t 
+  pretty (FunU ts t) = hsep $ punctuate " ->" (map prettyTypeU (ts <> [t]))
+  pretty (ForallU _ t) = pretty t
+  pretty t = prettyTypeU t
+
+prettyTypeU (ExistU v ts ds)
+  = angles $ pretty v
+  <> list (map prettyTypeU ts)
+  <> list (map prettyTypeU ds)
+prettyTypeU (ForallU _ t) = prettyTypeU t
+prettyTypeU (VarU (TV _ "Unit")) = "()"
+prettyTypeU (VarU v) = pretty v
+prettyTypeU (FunU [] t) = parens $ "<MISSING> -> " <> prettyTypeU t
+prettyTypeU (FunU ts t) = encloseSep "(" ")" " -> " (map prettyTypeU (ts <> [t]))
+prettyTypeU (AppU t ts) = hsep $ map parenTypeU (t:ts) where
+    parenTypeU t'@(AppU _ _) = parens $ prettyTypeU t'
+    parenTypeU t' = prettyTypeU t'
+prettyTypeU (NamU o n ps rs)
+    = block 4 (viaShow o <+> pretty n <> encloseSep "<" ">" "," (map pretty ps))
+              (vsep [pretty k <+> "::" <+> prettyTypeU x | (k, x) <- rs])
 
 instance Pretty UnresolvedPacker where
   pretty (UnresolvedPacker v t fs rs) = vsep